Move event logging script to shared folder
