+++ assets/setup-wizard/data/controls.js
@@ -7,7 +7,7 @@ import apiFetch from '@wordpress/api-fetch';
  * Internal dependencies
  */
 import { FETCH_FROM_API, APPLY_STEP_DATA } from './constants';
-import { logEvent } from '../log-event';
+import { logEvent } from '../../shared/helpers/log-event';
 
 export default {
 	/**
