+++ assets/setup-wizard/ready/mailinglist-signup-form.js
@@ -7,7 +7,7 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import { logLink } from '../log-event';
+import { logLink } from '../../shared/helpers/log-event';
 import { useSetupWizardStep } from '../data/use-setup-wizard-step';
 import List from '../../shared/components/list';
