+++ assets/shared/helpers/log-event.js
