+++ assets/shared/helpers/log-event.test.js
